// tirith-frontend/src/api/model/oAuth2AuthorizationCodeDto.ts
/**
 * Skribbl Typo API
 * Skribbl Typo API for resources, admin tools and authentification.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OAuth2ClientDto } from './oAuth2ClientDto';


export interface OAuth2AuthorizationCodeDto { 
    /**
     * The authorization code issued for the authenticated user and client
     */
    authorizationCode: string;
    client: OAuth2ClientDto;
}