// tirith-frontend/src/api/model/oAuth2AuthenticationResultDto.ts
/**
 * Skribbl Typo API
 * Skribbl Typo API for resources, admin tools and authentication.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OAuth2AuthorizationCodeDto } from './oAuth2AuthorizationCodeDto';
import { MemberDto } from './memberDto';


export interface OAuth2AuthenticationResultDto { 
    result: OAuth2AuthorizationCodeDto;
    member: MemberDto;
}